+++ src/AnomalyDetectors/IsolationForest.php
@@ -5,9 +5,10 @@ namespace Rubix\ML\AnomalyDetectors;
 use Rubix\ML\Learner;
 use Rubix\ML\Ensemble;
 use Rubix\ML\Persistable;
-use Rubix\ML\Probabilistic;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Datasets\Labeled;
+use Rubix\ML\Graph\Trees\ITree;
+use Rubix\ML\Other\Helpers\Stats;
 use InvalidArgumentException;
 use RuntimeException;
 
@@ -26,19 +27,8 @@ use RuntimeException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class IsolationForest implements Learner, Ensemble, Probabilistic, Persistable
+class IsolationForest implements Learner, Ensemble, Persistable
 {
-    const AVAILABLE_ESTIMATORS = [
-        IsolationTree::class,
-    ];
-
-    /**
-     * The base isolation tree to be used in the ensemble.
-     * 
-     * @var \Rubix\ML\Learner
-     */
-    protected $base;
-
     /**
      * The number of estimators to train in the ensemble.
      *
@@ -47,11 +37,26 @@ class IsolationForest implements Learner, Ensemble, Probabilistic, Persistable
     protected $estimators;
 
     /**
+     * The amount of contamination (outliers) that is presumed to be in
+     * the training set as a percentage.
+     *
+     * @var float
+     */
+    protected $contamination;
+
+    /**
      * The ratio of training samples to train each estimator on.
      *
      * @var float
      */
     protected $ratio;
+
+    /**
+     * The average depth of an isolation tree.
+     * 
+     * @var float|null
+     */
+    protected $pHat;
 
     /**
      * The trees that make up the forest.
@@ -63,26 +68,29 @@ class IsolationForest implements Learner, Ensemble, Probabilistic, Persistable
     ];
 
     /**
-     * @param  \Rubix\ML\Learner|null  $base
+     * The isolation score offset used by the decision function.
+     * 
+     * @var float|null
+     */
+    protected $offset;
+
+    /**
      * @param  int  $estimators
+     * @param  float  $contamination
      * @param  float  $ratio
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(?Learner $base = null, int $estimators = 100, float $ratio = 0.1)
+    public function __construct(int $estimators = 300, float $contamination = 0.1, float $ratio = 0.2)
     {
-        if (is_null($base)) {
-            $base = new IsolationTree();
-        }
-
-        if (!in_array(get_class($base), self::AVAILABLE_ESTIMATORS)) {
-            throw new InvalidArgumentException('Base estimator is not'
-                . ' compatible with this ensemble.');
-        }
-
         if ($estimators < 1) {
             throw new InvalidArgumentException("The number of estimators in the"
                 . " ensemble cannot be less than 1, $estimators given.");
+        }
+
+        if ($contamination < 0.) {
+            throw new InvalidArgumentException("Contamination cannot be less"
+                . " than 0, $contamination given.");
         }
 
         if ($ratio < 0.01 or $ratio > 1.) {
@@ -90,8 +98,8 @@ class IsolationForest implements Learner, Ensemble, Probabilistic, Persistable
                 . " 0.01 and 1, $ratio given.");
         }
 
-        $this->base = $base;
         $this->estimators = $estimators;
+        $this->contamination = $contamination;
         $this->ratio = $ratio;
     }
 
@@ -115,65 +123,98 @@ class IsolationForest implements Learner, Ensemble, Probabilistic, Persistable
      */
     public function train(Dataset $dataset) : void
     {
-        $p = (int) round($this->ratio * $dataset->numRows());
+        $n = $dataset->numRows();
+        
+        $maxDepth = (int) ceil(log(max($n, 2), 2));
+
+        $p = (int) round($this->ratio * $n);
+
+        $this->pHat = $this->c($p);
 
         $this->forest = [];
 
         for ($epoch = 0; $epoch < $this->estimators; $epoch++) {
-            $tree = clone $this->base;
+            $tree = new ITree($maxDepth);
 
             $subset = $dataset->randomize()->head($p);
 
-            $tree->train($subset);
+            $tree->grow($subset);
 
             $this->forest[] = $tree;
         }
+
+        $scores = [];
+
+        foreach ($dataset as $sample) {
+            $scores[] = $this->isolationScore($sample);
+        }
+
+        $shift = Stats::percentile($scores, 100. * $this->contamination);
+
+        $this->offset = $shift;
     }
 
     /**
      * Output a vector of class probabilities per sample.
      *
      * @param  \Rubix\ML\Datasets\Dataset  $dataset
+     * @throws \RuntimeException
      * @return array
      */
     public function predict(Dataset $dataset) : array
     {
+        if (is_null($this->offset)) {
+            throw new RuntimeException('Estimator has not been trained.');
+        }
+
         $predictions = [];
 
-        foreach ($this->proba($dataset) as $proba) {
-            $predictions[] = $proba > 0.5 ? 1 : 0;
+        foreach ($dataset as $sample) {
+            $score = $this->isolationScore($sample);
+
+            $score -= $this->offset;
+
+            $predictions[] = $score < 0. ? 1 : 0;
         }
 
         return $predictions;
     }
 
     /**
-     * Output a probability of being an anomaly.
-     *
-     * @param  \Rubix\ML\Datasets\Dataset  $dataset
-     * @throws \RuntimeException
-     * @return array
+     * Return the isolation score of a sample.
+     * 
+     * @param  array  $sample
+     * @return float
      */
-    public function proba(Dataset $dataset) : array
+    protected function isolationScore(array $sample) : float
     {
-        if (empty($this->forest)) {
-            throw new RuntimeException('Estimator has not been trained.');
-        }
-
-        $k = count($this->forest);
-
-        $probabilities = array_fill(0, $dataset->numRows(), 0.);
+        $depths = [];
 
         foreach ($this->forest as $tree) {
-            foreach ($tree->proba($dataset) as $i => $proba) {
-                $probabilities[$i] += $proba;
-            }
+            $node = $tree->search($sample);
+
+            $depths[] = isset($node) ? $node->depth() : self::EPSILON;
         }
 
-        foreach ($probabilities as &$proba) {
-            $proba /= $k;
-        }
+        $mean = Stats::mean($depths);
 
-        return $probabilities;
+        $score = 2. ** -($mean / $this->pHat);
+
+        return -$score;
+    }
+
+    /**
+     * Calculate the average path length of an unsuccessful search for n nodes.
+     *
+     * @param  int  $n
+     * @return float
+     */
+    protected function c(int $n) : float
+    {
+        if ($n <= 1) {
+            return 1.;
+        }
+        
+        return 2. * (log($n - 1) + M_EULER) - 2. * ($n - 1) / $n;
     }
 }