+++ src/Clusterers/FuzzyCMeans.php
@@ -274,7 +274,7 @@ class FuzzyCMeans implements Learner, Probabilistic, Persistable
      * @throws \RuntimeException
      * @return array
      */
-    public function initializeCentroids(Dataset $dataset) : array
+    protected function initializeCentroids(Dataset $dataset) : array
     {
         $n = $dataset->numRows();