+++ CHANGELOG.md
@@ -28,6 +28,7 @@
     - Added smart cluster initialization to K Means and Fuzzy C Means
     - Circle and Half Moon generators now generate Labeled datasets
     - Gaussian Mixture now uses K Means initialization
+    - Removed Isolation Tree anomaly detector
 
 - 0.0.5-alpha
     - Added Gaussian Mixture clusterer